+++ SmartPower/Core/Configurator.cs
@@ -6,49 +6,32 @@ using Bindito.Core;
 using HarmonyLib;
 using IgorZ.TimberDev.Utils;
 using Timberborn.MechanicalSystem;
-using Timberborn.PowerGenerating;
-using UnityDev.Utils.LogUtilsLite;
 using UnityEngine;
 
 // ReSharper disable InconsistentNaming
 // ReSharper disable UnusedMember.Local
 
-namespace IgorZ.SmartPower.Core {
+namespace IgorZ.SmartPower.Core;
 
 [Context("Game")]
 // ReSharper disable once UnusedType.Global
 sealed class Configurator : IConfigurator {
-  static readonly PrefabPatcher.RequiredComponentsDep PoweredAttractionDeps =
-      new(AccessTools.TypeByName("Timberborn.Attractions.AttractionSpec"), typeof(MechanicalBuilding));
-  static readonly PrefabPatcher.RequiredComponentsDep ManufactoryDeps =
-      new(AccessTools.TypeByName("Timberborn.Workshops.ManufactorySpec"), typeof(MechanicalBuilding));
-  static readonly PrefabPatcher.RequiredComponentsDep WalkerPoweredGeneratorDeps =
-      new(AccessTools.TypeByName("Timberborn.PowerGenerating.WalkerPoweredGenerator"));
+  static readonly PrefabPatcher.RequiredComponentsDep PoweredManufactoryDep =
+      new(typeof(MechanicalBuilding), AccessTools.TypeByName("Timberborn.Workshops.ManufactorySpec"));
+  static readonly PrefabPatcher.RequiredComponentsDep PoweredAttractionDep =
+      new(typeof(MechanicalBuilding), AccessTools.TypeByName("Timberborn.Attractions.AttractionSpec"));
   static readonly string PatchId = typeof(Configurator).AssemblyQualifiedName;
 
   public void Configure(IContainerDefinition containerDefinition) {
-    if (Features.DebugExVerboseLogging && DebugEx.LoggingSettings.VerbosityLevel < 5) {
-      DebugEx.LoggingSettings.VerbosityLevel = 5;
-    }
     HarmonyPatcher.PatchRepeated(
         PatchId + "-core",
-        typeof(MechanicalBuildingPatch), typeof(WalkerPoweredGeneratorPatch));
-
+        typeof(MechanicalBuildingPatch), typeof(MechanicalNodePatch));
+    containerDefinition.Bind<SmartPowerService>().AsSingleton();
     CustomizableInstantiator.AddPatcher(PatchId + "-instantiator", PatchMethod);
   }
 
   static void PatchMethod(GameObject prefab) {
-    PrefabPatcher.ReplaceComponent<GoodPoweredGenerator, SmartGoodPoweredGenerator>(prefab);
-    PrefabPatcher.AddComponent<SmartManufactory>(prefab, ManufactoryDeps.Check);
-    PrefabPatcher.AddComponent<SmartPoweredAttraction>(prefab, PoweredAttractionDeps.Check);
-
-    PrefabPatcher.AddComponent<PowerOutputBalancer>(
-        prefab, WalkerPoweredGeneratorDeps.Check, onAdd: balancer => {
-          balancer.runWhenPaused = true;
-          balancer.waitTicks = 3;
-        });
-    PrefabPatcher.AddComponent<AutoPausePowerGenerator>(prefab, WalkerPoweredGeneratorDeps.Check);
+    PrefabPatcher.AddComponent<SmartManufactory>(prefab, PoweredManufactoryDep.Check);
+    PrefabPatcher.AddComponent<SmartPoweredAttraction>(prefab, PoweredAttractionDep.Check);
   }
-}
-
 }